voicebot/pyproject.toml
[tool.poetry]
name = "voicebot"
version = "0.1.0"
description = ""
authors = ["bisd98 <107929119+bisd98@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
discord-py = {git = "https://github.com/Rapptz/discord.py"}
discord-ext-voice-recv = {git = "https://github.com/imayhaveborkedit/discord-ext-voice-recv"}
openai = "^1.13.3"
google-cloud-speech = "^2.25.1"
google-cloud-texttospeech = "^2.16.3"
google-api-core = "^2.17.1"
google-auth = "^2.28.2"
numpy = "^1.26.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
